HLE: Move svc.{cpp,h} to kernel/
